Made small changes like clarifying variable names, declaring some structs in header files, and replacing serial.print with log_

+++ SensorGridPM/config.h
@@ -12,7 +12,7 @@
 #include <KnightLab_SDConfig.h>
 #include <HONEYWELL_HPM.h>
 #include <KL_ADAFRUIT_SI7021.h>
-//#include "WatchdogSAMD.h"
+// #include "WatchdogSAMD.h"
 
 /**
  * SensorGrid will not print to serial if USB is not attached. This can be
@@ -93,7 +93,7 @@ struct Config {
     uint32_t heartbeat_period;  // in seconds
     uint32_t sample_period;  // in minutes
     uint32_t collection_period;  // in minutes
-    uint32_t SD_CHIP_SELECT_PIN;
+    uint8_t SD_CHIP_SELECT_PIN;
     uint32_t RFM95_CS;
     uint32_t RFM95_RST;
     uint32_t RFM95_INT;
@@ -132,7 +132,6 @@ struct SensorConfig {
     SensorReadFunction read_function;
     SensorStopFunction stop_function;
     struct SensorConfig *next;
-
 };
